apps/workers/requirements.txt
redis==5.0.1
asyncpg==0.29.0
sqlalchemy==2.0.23
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sentry-sdk==1.38.0
python-json-logger==2.0.7
tenacity==8.2.3
